requirements.txt
google-generativeai>=0.7.2
pinecone-client>=5.0.0
langchain>=0.2.10
langchain-community>=0.2.10
pypdf>=4.2.0
python-dotenv>=1.0.1
docx2txt>=0.8
tqdm>=4.66.4
